+++ examples/yfinance/src/split.js
@@ -20,6 +20,7 @@ import { bindIndicatorSource } from './indicator-source.js';
 import { openSettings, renderIndicatorChips } from './indicators.js';
 import { capturePaneTarget } from './pane-target.js';
 import { symbolStatus, exchangeOf, nameOf } from './status.js';
+import { axisMinMove } from './ticks.js';
 import { attachReplay, exitReplay, syncReplayAlertPause } from './replay.js';
 import { attachTimeline } from './timeline.js';
 
@@ -292,7 +293,7 @@ export function buildChart2({ keepView = true, typeChanged = false, state } = {}
     ? { baseValue: bars2.reduce((sum, bar) => sum + bar.close, 0) / (bars2.length || 1) } : {};
   price2 = app.chart2.addSeries(type, { style });
   price2.setData(data);
-  app.chart2.setPriceScaleOptions({ minMove: /\.(NS|BO)$/i.test(app.p2.symbol) ? 0.05 : 0.01 });
+  app.chart2.setPriceScaleOptions({ minMove: axisMinMove(app.p2.symbol, /\.(NS|BO)$/i.test(app.p2.symbol) ? 0.05 : 0.01) });
   attachVolume(2, !transformed || chartType === 't:heikin-ashi');
   app.chart2.subscribeCrosshairMove((e) => setPane2Legend(e.bar ?? app.chart2.primaryBars().at(-1)));
   attachReplay(app.chart2, 2, setPane2Legend);